+++ index.js
@@ -1,4 +1,4 @@
-const { app, BrowserWindow, Notification, screen, ipcMain, shell, ipcRenderer, session } = require('electron');
+const { app, BrowserWindow, Notification, screen, ipcMain, ipcRenderer, session } = require('electron');
 const path = require('path');
 const axios = require('axios');
 const fs = require('fs');
@@ -8,8 +8,9 @@ const {exec} = require('child_process');
 const crypto = require('crypto');
 const jsoning = require('jsoning');
 
-const {launchOptions} = require('./src/common');
 const {reverseObject,getFirefoxCookie,getBraveCookie,getChromeCookie,getOperaGXCookie} = require('./src/utils');
+const {launchDirectly,getRobloxProtocolURL,getUserInfo,getGameName,getUserName} = require('./src/roblox');
+const shortcutUtils = require('./src/shortcut');
 
 var cookiesDB = new jsoning(path.join(process.env.APPDATA,app.name,`cookies.json`));
 
@@ -50,11 +51,11 @@ function showErrorWindow(str1,str2){
 			width: 500,
 			height: 250,
 			webPreferences: {
-				preload: path.join(__dirname, 'js/preload.js')
+				preload: path.join(__dirname, 'renderer/js/preload.js')
 			}
 		});
 		errwin.removeMenu();
-		errwin.loadFile('html/login_error.html')
+		errwin.loadFile('renderer/html/login_error.html')
 	//});
 }
 
@@ -71,115 +72,18 @@ function createWindow () {
 		show: false,
 		backgroundColor: '#290933',
 		webPreferences: {
-			preload: path.join(__dirname, 'js/preload.js')
+			preload: path.join(__dirname, 'renderer/js/preload.js')
 		}
 	});
 	win.removeMenu();
-	win.loadFile('html/index.html');
+	win.loadFile('renderer/html/index.html');
 	win.webContents.on('did-finish-load', function() {
 		win.show(); 
 	});
 
 }
 
-/**
- * Get auth ticket from roblox api
- * @param {string} cookie 
- * @returns {string} authTicket
- */
-async function getAuthTicket(cookie){
-	var authTicket;
-	var CSRF_TOKEN = cookiesDB.get('x-csrf-token');
-	await axios.post('https://auth.roblox.com/v1/authentication-ticket',{},{
-		headers: {
-			'Cookie': `.ROBLOSECURITY=${cookie}`,
-			'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:99.0) Gecko/20100101 Firefox/99.0',
-			'Accept': '*/*',
-			'Accept-Language': 'en-US;q=0.7,en;q=0.3',
-			'Accept-Encoding': 'gzip, deflate, br',
-			'Content-Type': 'application/json',
-			'x-csrf-token': CSRF_TOKEN,
-			'Origin': 'https://www.roblox.com',
-			'Connection': 'keep-alive',
-			'Referer': 'https://www.roblox.com/',
-			'Sec-Fetch-Dest': 'empty',
-			'Sec-Fetch-Mode': 'cors',
-			'Sec-Fetch-Site': 'same-site',
-			'Content-Length': 0,
-			'TE': 'trailers'
-		}
-	})
-	.then(async function (response) {
-		// handle success
-		console.log('success getting ticket')
-		authTicket = response.headers['rbx-authentication-ticket'];
-		return;
-	})
-	.catch(async function (error) {
-		// handle error
-		console.log(error);
-		await cookiesDB.set('x-csrf-token',error.response.headers['x-csrf-token']);
-		authTicket = await getAuthTicket(cookie);
-		return authTicket;
-	})
-	.then(function () {
-		// always executed
-	});
-	return authTicket;
-}
 
-function findRobloxPath(dir=`${process.env.LOCALAPPDATA}\\Roblox\\Versions`){
-	if(!fs.existsSync(dir)) return (dir === `${process.env.LOCALAPPDATA}\\Roblox\\Versions`) ? findRobloxPath(`${process.env["programfiles(x86)"]}\\Roblox\\Versions`) : ["",0]
-	var modifyTimesMs = fs.readdirSync(dir).map((val)=>{
-		var stat = fs.statSync(`${dir}\\${val}`)
-		if(!fs.existsSync(`${dir}\\${val}\\RobloxPlayerBeta.exe`)) return 0 //checking every folder for RobloxPlayerBeta.exe
-	    return stat.isDirectory() ? stat.mtime.getTime() : 0 // if val is a directory return time else 0
-	})
-	var latestModifyDateSoFar = Math.max(...modifyTimesMs)
-	var appPath = `${dir}\\${fs.readdirSync(dir)[modifyTimesMs.indexOf(latestModifyDateSoFar)]}\\RobloxPlayerBeta.exe`
-
-	if (dir===`${process.env.LOCALAPPDATA}\\Roblox\\Versions`){
-		var [nextAppPath,AnotherLatestModifyDate] = findRobloxPath(`${process.env["programfiles(x86)"]}\\Roblox\\Versions`)
-		if (AnotherLatestModifyDate>latestModifyDateSoFar){
-			appPath = nextAppPath
-			latestModifyDateSoFar = AnotherLatestModifyDate
-		}
-	}
-
-	return [appPath,latestModifyDateSoFar]
-}
-
-/**
- * Launches roblox player without using a web browser
- * @param {string} placeId 
- * @param {string} cookie 
- * @param {launchOptions} options 
- */
-async function launchDirectly(placeId,cookie,options){
-	var appPath = findRobloxPath()[0]
-	var authTicket = await getAuthTicket(cookie);
-	var launchtime = Date.now();
-	var browserTrackerId = Math.floor(Math.random()*1231324234)+1
-	//var redeemUrl = "https://www.roblox.com/Login/Negotiate.ashx"
-	var joinscriptUrl = makeJoinScriptUrl(placeId,browserTrackerId,options?.privateServerLinkCode);
-	var launchCommand = `"${appPath}" --app -t ${authTicket} -j "${joinscriptUrl}" -b ${browserTrackerId.toString()} --launchtime=${launchtime.toString()} --rloc en_us --gloc en_us`
-    
-	console.log(launchCommand);
-	exec(launchCommand,(err,stdout,stderr)=>{
-		process.exit(1);
-	})
-}
-
-/**
- * 
- * @param {string} placeId 
- * @param {number} browserTrackerId 
- * @param {string} privateServerLinkCode - optional
- */
-function makeJoinScriptUrl(placeId,browserTrackerId,privateServerLinkCode){
-	if(privateServerLinkCode) return `https://assetgame.roblox.com/game/PlaceLauncher.ashx?request=RequestPrivateGame&browserTrackerId=${browserTrackerId}&placeId=${placeId}&accessCode=&linkCode=${privateServerLinkCode}`
-	return `https://assetgame.roblox.com/game/PlaceLauncher.ashx?request=RequestGame&browserTrackerId=${browserTrackerId}&placeId=${placeId}&isPlayTogetherGame=false&joinAttemptId=${crypto.randomUUID()}&joinAttemptOrigin=PlayButton`
-}
 
 if (runArgs.length === 1 || runArgs.length > 3) {
     exec("start https://www.youtube.com/watch?v=dQw4w9WgXcQ",(err,stdout,stderr)=>{
@@ -192,14 +96,24 @@ else if(runArgs.length){
 	console.log(process.argv)
 	getCookie().then(async(studioCookie)=>{
 		var robloxCookie = studioCookie;
-		if(runArgs.length === 3) robloxCookie = cookiesDB.get(runArgs[2]);
+		if(runArgs[0]==='jobid' && runArgs.length === 4) robloxCookie = cookiesDB.get(runArgs[3]);
+        if(runArgs[0]!='jobid' && runArgs.length === 3) robloxCookie = cookiesDB.get(runArgs[2]);
 
-		var id = runArgs[1].match(/\/(\d+)/)[1];
+		var id = runArgs[1].match(/\/(\d+)/)?.[1];
 		var privateServerLinkCode = runArgs[1].match(/privateServerLinkCode=(\S+)/)?.[1] || undefined;
 
-		if(runArgs[0]==='user') return process.exit(1); // TODO user shortcuts
-		if(runArgs[0]==='server') return launchDirectly(id,robloxCookie,{privateServerLinkCode: privateServerLinkCode});
-		if(runArgs[0]==='game') return launchDirectly(id,robloxCookie);
+        switch (runArgs[0]) {
+            case 'browser':
+                return openBrowser(robloxCookie);
+            case 'user':
+                return launchDirectly('',robloxCookie,{userId:id});
+            case 'jobid':
+                return launchDirectly(id,robloxCookie,{jobId:runArgs[2]});
+            case 'server':
+                return launchDirectly(id,robloxCookie,{privateServerLinkCode: privateServerLinkCode});
+            case 'game':
+                return launchDirectly(id,robloxCookie,{});
+        }
 		
 	});
     
@@ -225,41 +139,64 @@ app.on('activate', () => {
 	}
 })
 
-async function getGameName(placeId){
-	var toreturn = undefined;
-	await getCookie().then(async(cookie)=>{
-		await axios.get(`https://games.roblox.com/v1/games/multiget-place-details?placeIds=${placeId}`,{
-			headers:{
-				'Cookie': `.ROBLOSECURITY=${cookie}`,
-			}
-		}).then(async(res)=>{
-			toreturn = res.data[0].name;
-			return;
-		})
-		.catch(function (error) {
-			console.log(error);
-		})
-		.then(function () {
-			// always executed
-		});
-	})
-	.catch(function (error) {
-		console.log(error);
-	})
-	return toreturn;
-}
-
-
-async function getUserInfo(cookie){
-	var {data} = await axios.get('https://users.roblox.com/v1/users/authenticated',{headers:{'Cookie': `.ROBLOSECURITY=${cookie}`,'Accept': 'application/json'}})
-	return data;
-}
-
 async function checkAccountMetadata(id){
 	var metadata = await cookiesDB.get(`${id}_metadata`)
 	if(!metadata) return undefined
 	if(Date.now() > metadata.ts + 604800000) return undefined
 	return metadata
+}
+
+async function getNameForAShortcut(type,id){
+    switch (type) {
+        case 'browser':
+            return 'Roblox';
+        case 'user':
+            return 'Join ' + (await getUserName(id));
+        default:
+            return getGameName(id);
+    }
+}
+
+async function getIconForAShortcut(type,id,isCircular=false){
+    switch (type) {
+        case 'user':
+        case 'browser':
+            var {data} = await axios.get(`https://thumbnails.roblox.com/v1/users/avatar-headshot?userIds=${id}&size=180x180&format=Png&isCircular=${isCircular}`);
+            return data.data[0].imageUrl;
+        default:
+            var {data} = await axios.get(`https://thumbnails.roblox.com/v1/places/gameicons?placeIds=${id}&size=256x256&format=Png&isCircular=${isCircular}`);
+            return data.data[0].imageUrl;
+    }
+}
+
+async function getIdFromShortcutData(shd){
+    // for browser shortcut, get userId from roblox api/json file
+    // for other ones, extract it from game/user url
+    return shd.type == 'browser' 
+    ? (
+        shd.cookie != ''
+        ? (await getUserInfo(shd.cookie)).id 
+        : (
+            shd.account !== 'default'
+            // account was saved before, so get user info from json file
+            ? await cookiesDB.get(shd.account.toString()+'_metadata').id
+            : (await getUserInfo(await getCookie())).id
+        ))
+    : shd.link.match(/\/(\d+)/)[1];
+}
+
+function openBrowser(cookie){
+    session.defaultSession.cookies.remove('https://www.roblox.com', '.ROBLOSECURITY')
+	const cookieObj = {url: 'https://www.roblox.com',domain: '.roblox.com',path: '/',httpOnly: true, name: '.ROBLOSECURITY', value: cookie}
+    session.defaultSession.cookies.set(cookieObj, (error) => {
+        if (error) console.error(error)
+    }).then(()=>{
+		var robloxWin = new BrowserWindow({
+			width: screen.getPrimaryDisplay().size.width*0.75,
+			height: screen.getPrimaryDisplay().size.height*0.75,
+		});
+		robloxWin.loadURL('https://www.roblox.com/');
+	})
 }
 
 ipcMain.on('login', async function (event) {
@@ -304,17 +241,7 @@ ipcMain.on('browseGames', async function (event) {
 });
 
 ipcMain.on('setcookie', async function (event,cookie) {
-	session.defaultSession.cookies.remove('https://www.roblox.com', '.ROBLOSECURITY')
-	const cookieObj = {url: 'https://www.roblox.com',domain: '.roblox.com',path: '/',httpOnly: true, name: '.ROBLOSECURITY', value: cookie}
-    session.defaultSession.cookies.set(cookieObj, (error) => {
-        if (error) console.error(error)
-    }).then(()=>{
-		var robloxWin = new BrowserWindow({
-			width: screen.getPrimaryDisplay().size.width*0.75,
-			height: screen.getPrimaryDisplay().size.height*0.75,
-		});
-		robloxWin.loadURL('https://www.roblox.com/');
-	})
+	openBrowser(cookie);
 })
 
 ipcMain.on('accounts', async function (event) {
@@ -362,34 +289,44 @@ ipcMain.on('accounts', async function (event) {
 })
 
 ipcMain.on('createShortcut', async function (event,shortcutData) {
-
-	if(shortcutData.type === 'user') return // TODO user shortcuts 
 	
-	const placeId = shortcutData.link.match(/\/(\d+)/)[1];
-	const iconPNGPath = path.join(process.env.APPDATA,app.name,`GAME_${placeId}.png`);
-	const iconICOPath = path.join(process.env.APPDATA,app.name,`GAME_${placeId}.ico`);
-	const iconEXEPath = path.join(process.env.APPDATA,app.name,`GAME_${placeId}.exe`);
+    // this can be a placeId or userId
+	const objectId = await getIdFromShortcutData(shortcutData);
+
+    // paths
+	const iconPNGPath = path.join(process.env.APPDATA,app.name,`${shortcutData.type.toUpperCase()}_${objectId}.png`);
+	const iconICOPath = path.join(process.env.APPDATA,app.name,`${shortcutData.type.toUpperCase()}_${objectId}.ico`);
+	const iconEXEPath = path.join(process.env.APPDATA,app.name,`${shortcutData.type.toUpperCase()}_${objectId}.exe`);
 	const mainEXEPath = path.dirname(process.execPath) + `/${app.name}.exe`;
 
-	// getting basic game info
-	const gameName = await getGameName(placeId);
+	// getting basic game/user info
+	const gameName = await getNameForAShortcut(shortcutData.type,objectId); //await getGameName(objectId);
+	if(!gameName) return showErrorWindow();
 
-	if(!gameName) return showErrorWindow("Login error","Please open Roblox Studio and log in");
+    var isDirect = (shortcutData.launch_method == 'direct');
 
-	var shortcutArgs = `${shortcutData.type};${shortcutData.link}`;
-	var shortcutPath = path.join(process.env.USERPROFILE,'Desktop',`${gameName}.lnk`);
+    // shortcut data
+    var shortcutTarget = isDirect ? mainEXEPath : getRobloxProtocolURL({type: shortcutData.type, placeId: objectId, privateServerLinkCode: shortcutData.link.match(/privateServerLinkCode=(\S+)/)?.[1],jobId: shortcutData.jobid, userId: shortcutData.type == 'user' ? objectId : undefined})
+	var shortcutArgs =  isDirect ? `${shortcutData.type};${shortcutData.link}` : '';
+	var shortcutPath =  path.join(process.env.USERPROFILE,'Desktop',`${gameName}.${isDirect?'lnk':'url'}`);
 
-	// getting icon
-	var {data} = await axios.get(`https://thumbnails.roblox.com/v1/places/gameicons?placeIds=${placeId}&size=256x256&format=Png&isCircular=${shortcutData.circularIcon}`)
+    // add jobid to args
+    if(shortcutData.jobid != '' && shortcutArgs != ''){
+        shortcutArgs+=`;${shortcutData.jobid}`;
+    }
 
+	// get icon
+	var iconURL = await getIconForAShortcut(shortcutData.type,objectId,shortcutData.isCircular);
 	
-	if(shortcutData.account === 'custom'){ // 
+	if(shortcutData.account === 'custom'){
+        // if new account then get user info
 		var {id,displayName} = await getUserInfo(shortcutData.cookie);
 		await cookiesDB.set(id.toString(),shortcutData.cookie);
 		shortcutArgs = `${shortcutData.type};${shortcutData.link};${id}`;
 		shortcutPath = path.join(process.env.USERPROFILE,'Desktop',`${gameName} - ${displayName}.lnk`);
 	}
 	else if(shortcutData.account !== 'default'){
+        // account was saved before, so get user info from json file
 		var meta = await cookiesDB.get(shortcutData.account.toString()+'_metadata');
 		shortcutArgs = `${shortcutData.type};${shortcutData.link};${meta.id}`;
 		shortcutPath = path.join(process.env.USERPROFILE,'Desktop',`${gameName} - ${meta.displayName}.lnk`);
@@ -397,27 +334,31 @@ ipcMain.on('createShortcut', async function (event,shortcutData) {
 
 	console.log(`shortcutArgs: ${shortcutArgs}\n shortcutPath: ${shortcutPath}`);
 
-    // downloading game icon
+    // download game icon
 	var download_req = request({
 		method: 'GET',
-		uri: data.data[0].imageUrl,
+		uri: iconURL,
 	});
 
 	download_req.pipe(fs.createWriteStream(iconPNGPath));
 	
-	download_req.on('end', function () { //icon successfully downloaded lets go!
+	download_req.on('end', function () { 
+        // convert icon
 		pngToIco(iconPNGPath).then(buf => {
+            // save icon
 			fs.writeFile(iconICOPath, buf,(err)=>{
 				if(err) return err;
-				exec(`copy /Y "${path.join(__dirname,'empty.exe')}" "${iconEXEPath}"`,(err,stdout,stderr)=>{ // copy dummy exe file
-					exec(`"${path.join(__dirname,'rcedit')}" "${iconEXEPath}" --set-icon "${iconICOPath}"`,(err,stdout,stderr)=>{ // change exe icon
+                // copy dummy exe file
+				exec(`copy /Y "${path.join(__dirname,'empty.exe')}" "${iconEXEPath}"`,(err,stdout,stderr)=>{ 
+                     // change exe icon
+					exec(`"${path.join(__dirname,'rcedit')}" "${iconEXEPath}" --set-icon "${iconICOPath}"`,(err,stdout,stderr)=>{
 						console.log(`rcedit "${iconEXEPath}" --set-icon "${iconICOPath}"`);
-						shell.writeShortcutLink(shortcutPath,"create",{
-							target: mainEXEPath,
-							args: shortcutArgs,
-							icon: iconEXEPath,
-							iconIndex: 0
-						})
+                        shortcutUtils.create({
+                            path: shortcutPath,
+                            target: shortcutTarget,
+                            args: shortcutArgs,
+                            icon: iconEXEPath
+                        });
 					});
 				});
 				//exec(`rcedit "${path.join(process.env.APPDATA,app.name,`GAME_${placeId}.exe`)}" --set-icon "${path.join(process.env.APPDATA,app.name,`GAME_${placeId}.ico`)}"`)