+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## rlnk v1.0.3
+
+Release Notes:
+- 4 new types of shortcuts:
+  - user shortcut
+  - jobId shortcut
+  - browser shortcut
+  - app shortcut
+- splitted the code across multiple files, to make it easier to read
+
+## rlnk v1.0.2
+
+Release Notes:
+- Added support for the `roblox://` protocol launching
+
 ## rlnk v1.0.1
 
 Release Notes: